Custom Energy Zones

+++ CO2Toolkit/Zone.cs
@@ -4,13 +4,13 @@
     {
         public string ThreeLetterCode { get; }
         public string Name { get; }
-        public float AverageCarbonGridIntensity { get; }
+        public double AverageCarbonGridIntensity { get; }
         public uint Year { get; }
 
         public Zone(
             string threeLetterCode,
             string name,
-            float avgCarbonGridIntensity,
+            double avgCarbonGridIntensity,
             uint year)
         {
             ThreeLetterCode = threeLetterCode;
